Removed AuthManager
